chore - restructure character/characters
Update return type for list characters as well
