+++ apps/backend/src/character/character.module.ts
@@ -1,0 +1,26 @@
+import { Module } from '@nestjs/common';
+import { PrismaModule } from '../prisma/prisma.module';
+import { SupabaseModule } from '../supabase/supabase.module';
+import { OpenAiModule } from '../openai/openai.module';
+import { CharacterController } from './character.controller';
+import { CreateCharacterSuggestionService } from './services/create-character-suggestion.service';
+import { CreateCharacterService } from './services/create-character.service';
+import { FILE_STORAGE } from '../common/tokens';
+import { MockFileStorage } from '../common/storage/mock-file-storage.service';
+import { SupabaseFileStorage } from '../common/storage/supabase-file-storage.service';
+
+const useMockServices = process.env.USE_MOCK_SERVICES === 'true';
+
+@Module({
+  imports: [PrismaModule, SupabaseModule, OpenAiModule],
+  controllers: [CharacterController],
+  providers: [
+    CreateCharacterSuggestionService,
+    CreateCharacterService,
+    {
+      provide: FILE_STORAGE,
+      useClass: useMockServices ? MockFileStorage : SupabaseFileStorage, // Replace later
+    },
+  ],
+})
+export class CharacterModule {}
