+++ apps/backend/src/character/mappers/character.mapper.ts
@@ -1,5 +1,5 @@
 import { Prisma } from '@prisma/client';
-import { CharacterDto, StatType } from '../dto/character.dto';
+import { Character, StatType } from '../../common/types/character.types';
 
 export type FullCharacter = Prisma.CharacterGetPayload<{
   include: {
@@ -7,14 +7,17 @@ export type FullCharacter = Prisma.CharacterGetPayload<{
   };
 }>;
 
-export function toCharacterDto(character: FullCharacter): CharacterDto {
+export function toCharacterDto(character: FullCharacter): Character {
   return {
     id: character.id,
     name: character.name,
     description: character.description,
     lore: character.lore,
-    status: character.status,
     stats: character.stats as Record<string, number>,
+    status: character.status,
+    imageFrontUrl: character.imageFrontUrl ?? undefined,
+    imageBackUrl: character.imageBackUrl ?? undefined,
+    imageProfileUrl: character.imageProfileUrl ?? undefined,
     moves: character.moves.map((move) => ({
       name: move.name,
       description: move.description,
