+++ apps/backend/src/character/services/create-character-suggestion.service.ts
@@ -1,0 +1,16 @@
+import { Injectable } from '@nestjs/common';
+import { GenerateCharacterSuggestionsService } from '../../openai/queries/generate-character-suggestions.service';
+import { CreateCharacterSuggestionRequestDto } from '../dto/create-chracter-suggestion.request.dto';
+import { CharacterSuggestion } from '../../common/types/character.types';
+
+@Injectable()
+export class CreateCharacterSuggestionService {
+  constructor(
+    private readonly characterSuggestions: GenerateCharacterSuggestionsService,
+  ) {}
+  async execute(
+    dto: CreateCharacterSuggestionRequestDto,
+  ): Promise<CharacterSuggestion> {
+    return this.characterSuggestions.execute(dto.name, dto.description);
+  }
+}
