--- apps/backend/src/characters/dto/character-list.dto.ts
@@ -1,8 +1,0 @@
-import { CharacterStatus } from '@prisma/client';
-
-export class CharacterListItemDto {
-  id: string;
-  name: string;
-  imageProfileUrl: string | null;
-  status: CharacterStatus;
-}
